exit_mna_tool/requirements.txt
requests>=2.32.0
beautifulsoup4>=4.12.0
pydantic>=2.7.0
tldextract>=5.1.0
python-dateutil>=2.9.0
rapidfuzz>=3.9.0
typer>=0.12.0
rich>=13.7.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
openai>=1.0.0
flask>=2.0.0
readability-lxml>=0.8.1
flask-cors>=3.0.0